Moved adapter_for_test file to active_model_serializers folder and changed namespace of adapter inside the test file
